# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/channels/me": {
      "get": {
        "operationId": "ChannelsController_getMyChannel",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get authenticated user channel",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "operationId": "ChannelsController_updateMyChannel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update authenticated user channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/channels/me/videos": {
      "get": {
        "operationId": "ChannelsController_listMyVideos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all videos for owner channel dashboard",
        "tags": [
          "channels"
        ]
      }
    },
    "/channels/me/subscriptions": {
      "get": {
        "operationId": "ChannelsController_listMySubscriptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List channels the user subscribes to",
        "tags": [
          "channels"
        ]
      }
    },
    "/channels/{nickname}": {
      "get": {
        "operationId": "ChannelsController_getPublicChannel",
        "parameters": [
          {
            "name": "nickname",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Public channel profile and published videos",
        "tags": [
          "channels"
        ]
      }
    },
    "/channels/{nickname}/subscribe": {
      "post": {
        "operationId": "ChannelsController_subscribe",
        "parameters": [
          {
            "name": "nickname",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Subscribe to a channel",
        "tags": [
          "channels"
        ]
      },
      "delete": {
        "operationId": "ChannelsController_unsubscribe",
        "parameters": [
          {
            "name": "nickname",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unsubscribe from a channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/videos/{publicId}/comments": {
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List comments for a video",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a top-level comment",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{commentId}/replies": {
      "post": {
        "operationId": "CommentsController_reply",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reply to a comment",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{commentId}": {
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete own comment",
        "tags": [
          "comments"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateChannelDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}